// frontend/src/types/InGameState.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { ChatMessage } from "./ChatMessage";
import type { Curve } from "./Curve";
import type { Player } from "./Player";
import type { Word } from "./Word";

export type InGameState = { players: Array<Player>, current_player_index: number, curves: Array<Curve>, current_curve: Curve | null, current_round: number, max_rounds: number, word: Word, fake_artist: Player, spectators: Array<Player>, chat: Array<ChatMessage>, votes: { [key in string]?: string }, };