// frontend/src/types/JoinEvent.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { Game } from "./Game";
import type { Player } from "./Player";

export type JoinEvent = { current_player_id: string | null, game_state: Game | null, players: Array<Player> | null, spectators: Array<Player> | null, };